refactor(work-control): satisfy maintainability ratchet
