+++ extensions/mission-control/mission-control-actions/domain/__init__.py
@@ -2,6 +2,6 @@
 
 """Work Control Suite domain with explicit legacy compatibility readers."""
 
-from . import mission_bundle, mission_control
+from . import mission_bundle, mission_control, work_control
 
-__all__ = ["mission_bundle", "mission_control"]
+__all__ = ["mission_bundle", "mission_control", "work_control"]
