+++ package.json
@@ -129,7 +129,7 @@
     "check:agent-pack": "node scripts/require-shifu.mjs check:agent-pack && node scripts/verify-agent-pack.mjs",
     "check:fact-cut-kernel-contract": "node scripts/require-shifu.mjs check:fact-cut-kernel-contract && node --test scripts/check-fact-cut-kernel-contract.test.mjs",
     "check:work-lifecycle-operation-matrix": "node scripts/require-shifu.mjs check:work-lifecycle-operation-matrix && node scripts/materialize-work-lifecycle-operation-matrix.mjs --check && node scripts/render-work-lifecycle-operation-matrix.mjs --check && node developer/sdk/src/sdk.js contract render work-lifecycle-operation-matrix --check --json && node developer/sdk/src/sdk.js contract policy --check --json && node developer/sdk/src/sdk.js contract audit --json && node --test scripts/check-work-lifecycle-operation-matrix.test.mjs scripts/check-native-authority-membrane.test.mjs",
-    "check:work-control-vocabulary": "node scripts/require-shifu.mjs check:work-control-vocabulary && node scripts/check-work-control-vocabulary.mjs && node --test scripts/check-work-control-vocabulary.test.mjs",
+    "check:work-control-vocabulary": "node scripts/require-shifu.mjs check:work-control-vocabulary && node --test scripts/check-work-control-vocabulary.test.mjs",
     "render:work-lifecycle-operation-matrix": "node scripts/require-shifu.mjs render:work-lifecycle-operation-matrix && node scripts/materialize-work-lifecycle-operation-matrix.mjs --write && node scripts/render-work-lifecycle-operation-matrix.mjs --write && node developer/sdk/src/sdk.js contract render work-lifecycle-operation-matrix --write --json && node developer/sdk/src/sdk.js contract policy --write --json && node scripts/render-work-lifecycle-operation-matrix.mjs --sync-artifacts",
     "check:domain-profile-authoring": "node scripts/require-shifu.mjs check:domain-profile-authoring && node scripts/domain-profile-authoring.mjs --check && node developer/sdk/src/sdk.js contract render domain-profile-authoring --check --json && node developer/sdk/src/sdk.js contract policy --check --json && node developer/sdk/src/sdk.js contract audit --json && node --test scripts/check-domain-profile-authoring.test.mjs",
     "render:domain-profile-authoring": "node scripts/require-shifu.mjs render:domain-profile-authoring && node scripts/domain-profile-authoring.mjs --write && node developer/sdk/src/sdk.js contract render domain-profile-authoring --write --json && node developer/sdk/src/sdk.js contract policy --write --json",
